Move circuit compiler (aka factory) to separate package
