+++ pkg/policies/controlplane/policy.go
@@ -18,6 +18,7 @@ import (
 	"github.com/fluxninja/aperture/pkg/jobs"
 	"github.com/fluxninja/aperture/pkg/log"
 	"github.com/fluxninja/aperture/pkg/notifiers"
+	"github.com/fluxninja/aperture/pkg/policies/controlplane/circuitcompiler"
 	"github.com/fluxninja/aperture/pkg/policies/controlplane/components"
 	"github.com/fluxninja/aperture/pkg/policies/controlplane/iface"
 	"github.com/fluxninja/aperture/pkg/policies/controlplane/resources/classifier"
@@ -32,7 +33,7 @@ func policyModule() fx.Option {
 	componentFactoryOptions := components.FactoryModule()
 
 	return fx.Options(
-		circuitFactoryModule(),
+		circuitcompiler.Module(),
 		componentFactoryOptions,
 	)
 }
@@ -92,7 +93,7 @@ func newPolicyOptions(
 }
 
 // CompilePolicy takes policyMessage and returns a compiled policy. This is a helper method for standalone consumption of policy compiler.
-func CompilePolicy(policyMessage *policylangv1.Policy, registry status.Registry) (CompiledCircuit, error) {
+func CompilePolicy(policyMessage *policylangv1.Policy, registry status.Registry) (circuitcompiler.Circuit, error) {
 	wrapperMessage, err := hashAndPolicyWrap(policyMessage, "DoesNotMatter")
 	if err != nil {
 		return nil, err
@@ -105,7 +106,7 @@ func CompilePolicy(policyMessage *policylangv1.Policy, registry status.Registry)
 }
 
 // compilePolicyWrapper takes policyProto and returns a compiled policy.
-func compilePolicyWrapper(wrapperMessage *policysyncv1.PolicyWrapper, registry status.Registry) (*Policy, CompiledCircuit, fx.Option, error) {
+func compilePolicyWrapper(wrapperMessage *policysyncv1.PolicyWrapper, registry status.Registry) (*Policy, circuitcompiler.Circuit, fx.Option, error) {
 	if wrapperMessage == nil {
 		return nil, nil, nil, fmt.Errorf("nil policy wrapper message")
 	}
@@ -140,7 +141,7 @@ func compilePolicyWrapper(wrapperMessage *policysyncv1.PolicyWrapper, registry s
 			resourceOptions = append(resourceOptions, classifierOption)
 		}
 	}
-	var compiledCircuit CompiledCircuit
+	var compiledCircuit circuitcompiler.Circuit
 	partialCircuitOption := fx.Options()
 	var err error
 
@@ -148,7 +149,7 @@ func compilePolicyWrapper(wrapperMessage *policysyncv1.PolicyWrapper, registry s
 		// Read evaluation interval
 		policy.evaluationInterval = policyProto.GetCircuit().GetEvaluationInterval().AsDuration()
 
-		compiledCircuit, partialCircuitOption, err = compileCircuit(policyProto.GetCircuit().Components, policy)
+		compiledCircuit, partialCircuitOption, err = circuitcompiler.Compile(policyProto.GetCircuit().Components, policy)
 		if err != nil {
 			return nil, nil, nil, err
 		}
